pyproject.toml
[tool.poetry]
name = "finance-manager-dev"
version = "0.1.0"
description = ""
authors = ["Afonso Caboz <afonso.caboz@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.2"
django = "^5.0.4"
python-dateutil = "^2.9.0.post0"
plotly = "^5.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
